move flow definition page, lint fixes
