+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page.tsx
@@ -1,8 +1,8 @@
-import type { NextPageWithLayout } from "../../shared/layout";
-import { getLayoutWithSidebar } from "../../shared/layout";
-import { FlowDefinitionsContextProvider } from "../shared/flow-definitions-context";
-import { FlowRunsContextProvider } from "../shared/flow-runs-context";
-import { FlowDefinition } from "./flows.page/flow-definition";
+import type { NextPageWithLayout } from "../../../shared/layout";
+import { getLayoutWithSidebar } from "../../../shared/layout";
+import { FlowDefinitionsContextProvider } from "../../shared/flow-definitions-context";
+import { FlowRunsContextProvider } from "../../shared/flow-runs-context";
+import { FlowDefinition } from "./[flow-uuid].page/flow-definition";
 
 const FlowsDefinitionPage: NextPageWithLayout = () => {
   return (
