+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page/flow-definition.tsx
@@ -20,14 +20,14 @@ import type {
 } from "../../../graphql/api-types.gen";
 import { startFlowMutation } from "../../../graphql/queries/knowledge/entity.queries";
 import { HEADER_HEIGHT } from "../../../shared/layout/layout-with-header/page-header";
+import { useFlowDefinitionsContext } from "../../shared/flow-definitions-context";
+import { useFlowRunsContext } from "../../shared/flow-runs-context";
 import { ActivityLog } from "./flow-definition/activity-log";
 import { FlowRunSidebar } from "./flow-definition/flow-run-sidebar";
 import { Outputs } from "./flow-definition/outputs";
 import { RunFlowModal } from "./flow-definition/run-flow-modal";
 import { SectionLabel } from "./flow-definition/section-label";
 import { nodeDimensions } from "./flow-definition/shared/dimensions";
-import { useFlowDefinitionsContext } from "../../shared/flow-definitions-context";
-import { useFlowRunsContext } from "../../shared/flow-runs-context";
 import {
   flowSectionBorderRadius,
   transitionOptions,
