+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page/flow-definition/flow-run-sidebar.tsx
@@ -19,15 +19,15 @@ import { differenceInMilliseconds, intervalToDuration } from "date-fns";
 import type { PropsWithChildren } from "react";
 import { useEffect, useMemo, useState } from "react";
 
-import { isNonNullable } from "../../../../lib/typeguards";
-import { ArrowRightIcon } from "../../../../shared/icons/arrow-right";
-import { EllipsisRegularIcon } from "../../../../shared/icons/ellipsis-regular-icon";
-import { SectionLabel } from "./section-label";
+import { isNonNullable } from "../../../../../lib/typeguards";
+import { ArrowRightIcon } from "../../../../../shared/icons/arrow-right";
+import { EllipsisRegularIcon } from "../../../../../shared/icons/ellipsis-regular-icon";
 import {
   statusToSimpleStatus,
   useStatusForStep,
   useStatusForSteps,
-} from "../../../shared/flow-runs-context";
+} from "../../../../shared/flow-runs-context";
+import { SectionLabel } from "./section-label";
 import { flowSectionBorderRadius } from "./shared/styles";
 import type {
   FlowMaybeGrouped,
