+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page/flow-definition/outputs/shared/empty-output-box.tsx
