+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page/flow-definition/run-flow-modal.tsx
@@ -11,11 +11,11 @@ import { Box, Stack, Typography } from "@mui/material";
 import type { PropsWithChildren } from "react";
 import { useState } from "react";
 
-import { XMarkRegularIcon } from "../../../../shared/icons/x-mark-regular-icon";
-import { Button } from "../../../../shared/ui/button";
-import { Modal } from "../../../../shared/ui/modal";
-import { useAuthenticatedUser } from "../../../shared/auth-info-context";
-import { GoogleAuthProvider } from "../../../shared/integrations/google/google-auth-context";
+import { XMarkRegularIcon } from "../../../../../shared/icons/x-mark-regular-icon";
+import { Button } from "../../../../../shared/ui/button";
+import { Modal } from "../../../../../shared/ui/modal";
+import { useAuthenticatedUser } from "../../../../shared/auth-info-context";
+import { GoogleAuthProvider } from "../../../../shared/integrations/google/google-auth-context";
 import { ManualTriggerInput } from "./run-flow-modal/manual-trigger-input";
 import { WebSelector } from "./run-flow-modal/manual-trigger-input/web-selector";
 import type { FormState, LocalPayload } from "./run-flow-modal/types";
