+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page/flow-definition/shared/dimensions.ts
