+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page/flow-definition/shared/types.ts
@@ -8,7 +8,7 @@ import type {
 } from "@local/hash-isomorphic-utils/flows/types";
 import type { Edge, Node } from "reactflow";
 
-import type { SimpleStatus } from "../../../../shared/flow-runs-context";
+import type { SimpleStatus } from "../../../../../shared/flow-runs-context";
 
 export type NodeData = {
   kind: StepDefinition["kind"];
