+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page/flow-definition/swimlane.tsx
@@ -14,12 +14,12 @@ import ReactFlow, {
   useReactFlow,
 } from "reactflow";
 
-import { MarkerEnd } from "./marker-end";
-import { nodeTabHeight, parentGroupPadding } from "./shared/dimensions";
 import {
   useFlowRunsContext,
   useStatusForSteps,
-} from "../../../shared/flow-runs-context";
+} from "../../../../shared/flow-runs-context";
+import { MarkerEnd } from "./marker-end";
+import { nodeTabHeight, parentGroupPadding } from "./shared/dimensions";
 import { transitionOptions } from "./shared/styles";
 import type {
   CustomNodeType,
