+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page/flow-definition/swimlane/custom-node/node-container.tsx
@@ -3,9 +3,9 @@ import { Box, Stack, Typography } from "@mui/material";
 import type { PropsWithChildren } from "react";
 import { useNodeId, useStore } from "reactflow";
 
+import type { SimpleStatus } from "../../../../../../shared/flow-runs-context";
+import { useFlowRunsContext } from "../../../../../../shared/flow-runs-context";
 import { nodeTabHeight } from "../../shared/dimensions";
-import type { SimpleStatus } from "../../../../../shared/flow-runs-context";
-import { useFlowRunsContext } from "../../../../../shared/flow-runs-context";
 import { transitionOptions } from "../../shared/styles";
 import { statusSx } from "./node-styles";
