+++ apps/hash-frontend/src/pages/[shortname]/flows/[flow-uuid].page/flow-definition/swimlane/shared/edge-styles.ts
@@ -1,6 +1,6 @@
 import { customColors } from "@hashintel/design-system/theme";
 
-import type { SimpleStatus } from "../../../../../shared/flow-runs-context";
+import type { SimpleStatus } from "../../../../../../shared/flow-runs-context";
 
 export const edgeColor = {
   Complete: customColors.green[70],
