+++ libs/@hashintel/design-system/src/components.ts
@@ -13,8 +13,6 @@ export * from "./form-inline";
 export * from "./icon-angle-right-regular";
 export * from "./icon-arrow-left";
 export * from "./icon-arrow-right";
-export * from "./icon-bullseye-light";
-export * from "./icon-terminal-light";
 export * from "./icon-arrow-rotate-left";
 export * from "./icon-arrow-up-regular";
 export * from "./icon-arrow-up-right";
@@ -23,6 +21,7 @@ export * from "./icon-arrow-up-wide-short-light";
 export * from "./icon-asterisk-regular";
 export * from "./icon-at-regular";
 export * from "./icon-barcode";
+export * from "./icon-bullseye-light";
 export * from "./icon-button";
 export * from "./icon-caret-down-solid";
 export * from "./icon-check";
@@ -58,6 +57,7 @@ export * from "./icon-rotate-regular";
 export * from "./icon-ruler-regular";
 export * from "./icon-sparkles-light";
 export * from "./icon-table-light";
+export * from "./icon-terminal-light";
 export * from "./icon-wand-magic-sparkles";
 export * from "./image-with-checked-background";
 export * from "./input-props";
